refactor tests to use media in dataplane bucket, big test overhaul, small IAM changes for dataplane
