+++ source/dataplaneapi/external_resources.json
@@ -41,7 +41,8 @@
                     "s3:DeleteObject",
                     "s3:HeadBucket",
                     "s3:PutObjectAcl",
-                    "s3:GetObjectVersion"
+                    "s3:GetObjectVersion",
+                    "s3:DeleteObjectVersion"
                   ],
                   "Resource": {"Fn::Sub": "arn:aws:s3:::${DataplaneBucketName}/*"}
                 },
