+++ tests/deploy-transcribe.sh
