+++ tests/test_dataplane_crud.py
@@ -2,7 +2,7 @@
 # SPDX-License-Identifier: Apache-2.0
 
 ###############################################################################
-# Integration testing for the MIE workflow API
+# Integration testing for the MIE dataplane API
 #
 # PRECONDITIONS:
 # MIE base stack must be deployed in your AWS account
@@ -10,8 +10,8 @@
 # Boto3 will raise a deprecation warning (known issue). It's safe to ignore.
 #
 # USAGE:
-#   cd tests/
-#   pytest -s -W ignore::DeprecationWarning -p no:cacheprovider
+#   ./run_tests.sh from the tests dir
+#   
 #
 ###############################################################################
 
@@ -28,30 +28,46 @@ import re
 import os
 from jsonschema import validate
 
-# local imports
-import api
-import validation
+# Testing data
 
-REGION = os.environ['REGION']
-BUCKET_NAME = os.environ['BUCKET_NAME']
-MIE_STACK_NAME = os.environ['MIE_STACK_NAME']
-VIDEO_FILENAME = os.environ['VIDEO_FILENAME']
-IMAGE_FILENAME = os.environ['IMAGE_FILENAME']
-AUDIO_FILENAME = os.environ['AUDIO_FILENAME']
-TEXT_FILENAME = os.environ['TEXT_FILENAME']
+session_nonpaginated_results = {
+        "OperatorName": "samplenonpagresults",
+        "WorkflowId": "test-123",
+        "Results": {
+            "Testing": "This is some test data"
+        }
+    }
+
+
+session_paginated_results = {
+        "OperatorName": "samplepagresults",
+        "WorkflowId": "test-123",
+        "Results": {
+            "Labels": [{
+                "Timestamp": 0,
+                "Label": {
+                    "Name": "Purple",
+                    "Confidence": 54.41469192504883,
+                    "Instances": [],
+                    "Parents": []
+                }
+            }
+            ]
+        }
+    }
 
 # TODO: Add assert statements for status == success in api response json
 
 
-def test_dataplane_api(stack_resources, session_nonpaginated_results, session_paginated_results):
-
+def test_dataplane_api(api):
+    api = api()
     urllib3.disable_warnings(urllib3.exceptions.InsecureRequestWarning)
 
     # Create an asset
 
     print("Creating an asset in the dataplane")
 
-    create_asset_response = api.create_asset(stack_resources, BUCKET_NAME, VIDEO_FILENAME)
+    create_asset_response = api.create_asset()
     assert create_asset_response.status_code == 200
     asset = create_asset_response.json()
     asset_id = asset["AssetId"] 
@@ -62,7 +78,7 @@ def test_dataplane_api(stack_resources, session_nonpaginated_results, session_pa
 
     print("Adding nonpaginated metadata to asset: {asset}".format(asset=asset_id))
 
-    nonpaginated_post_response = api.post_metadata(stack_resources, asset_id, session_nonpaginated_results)
+    nonpaginated_post_response = api.post_metadata(asset_id, session_nonpaginated_results)
     assert nonpaginated_post_response.status_code == 200
     nonpaginated_post_results = nonpaginated_post_response.json()
     print("Successfully stored nonpaginated results for: {asset}".format(asset=asset_id))
@@ -72,12 +88,12 @@ def test_dataplane_api(stack_resources, session_nonpaginated_results, session_pa
 
     pages_stored = 0
     while 2 > pages_stored:
-        store_page_response = api.post_metadata(stack_resources, asset_id, session_paginated_results, paginate=True)
+        store_page_response = api.post_metadata(asset_id, session_paginated_results, paginate=True)
         assert store_page_response.status_code == 200
         print("Successfully stored a page of results")
         pages_stored += 1
 
-    paginated_post_response = api.post_metadata(stack_resources, asset_id, session_paginated_results, paginate=True,
+    paginated_post_response = api.post_metadata(asset_id, session_paginated_results, paginate=True,
                                                 end=True)
     assert paginated_post_response.status_code == 200
     paginated_post_results = paginated_post_response.json()
@@ -92,7 +108,7 @@ def test_dataplane_api(stack_resources, session_nonpaginated_results, session_pa
 
     more_results = True
     while more_results:
-        retrieve_metadata_response = api.get_all_metadata(stack_resources, asset_id, cursor)
+        retrieve_metadata_response = api.get_all_metadata(asset_id, cursor)
         assert retrieve_metadata_response.status_code == 200
         retrieved_metadata = retrieve_metadata_response.json()
         print(retrieved_metadata)
@@ -106,7 +122,7 @@ def test_dataplane_api(stack_resources, session_nonpaginated_results, session_pa
 
     print("Retrieving sample metadata for the asset: {asset}".format(asset=asset_id))
 
-    retrieve_single_metadata_response = api.get_single_metadata_field(stack_resources, asset_id,
+    retrieve_single_metadata_response = api.get_single_metadata_field(asset_id,
                                                                       session_nonpaginated_results)
     assert retrieve_single_metadata_response.status_code == 200
 
@@ -119,7 +135,7 @@ def test_dataplane_api(stack_resources, session_nonpaginated_results, session_pa
     # Delete specific metadata
 
     print("Deleting specific metadata for the asset: {asset}".format(asset=asset_id))
-    delete_metadata_response = api.delete_single_metadata_field(stack_resources, asset_id, session_nonpaginated_results)
+    delete_metadata_response = api.delete_single_metadata_field(asset_id, session_nonpaginated_results)
     assert delete_metadata_response.status_code == 200
     deleted_metadata = delete_metadata_response.json()
     print("Successfully deleted metadata field: {operator} for asset: {asset}".format(
@@ -129,7 +145,7 @@ def test_dataplane_api(stack_resources, session_nonpaginated_results, session_pa
     # Delete entire asset
 
     print("Deleting the asset from the dataplane")
-    delete_asset_response = api.delete_asset(stack_resources, asset_id)
+    delete_asset_response = api.delete_asset(asset_id)
     assert delete_asset_response.status_code == 200
     deleted_asset = delete_asset_response.text
     print(deleted_asset)
