+++ tests/test_instant_translate.py
@@ -28,53 +28,40 @@ import re
 import os
 from jsonschema import validate
 
-# local imports
-import api 
-import validation
 
-REGION = os.environ['REGION']
-BUCKET_NAME = os.environ['BUCKET_NAME']
-MIE_STACK_NAME = os.environ['MIE_STACK_NAME']
-VIDEO_FILENAME = os.environ['VIDEO_FILENAME']
-VIDEO_WITH_AUDIO_FILENAME=os.environ['VIDEO_WITH_AUDIO_FILENAME']
-IMAGE_FILENAME = os.environ['IMAGE_FILENAME']
-AUDIO_FILENAME = os.environ['AUDIO_FILENAME']
-TEXT_FILENAME = os.environ['TEXT_FILENAME']
-token = os.environ["MIE_ACCESS_TOKEN"]
-
-def test_instant_translate(stack_resources, api_schema):
-
+def test_instant_translate(api, testing_env_variables, stack_resources):
+    api = api()
     urllib3.disable_warnings(urllib3.exceptions.InsecureRequestWarning)
 
     # InstantTranslate workflow must be deployed
     
-    print("Testing InstantTranslate workflow.  BUCKET_NAME = "+BUCKET_NAME+" VIDEO_WITH_AUDIO_FILENAME = "+ VIDEO_WITH_AUDIO_FILENAME)
+    print("Testing InstantTranslate workflow.  BUCKET_NAME = "+testing_env_variables['BUCKET_NAME']+" SAMPLE_VIDEO = "+ testing_env_variables['SAMPLE_VIDEO'])
 
     body = {
         "Name":"InstantTranslateWorkflow",
         "Input": {
             "Media": {
                 "Video": {
-                    "S3Bucket": BUCKET_NAME,
-                    "S3Key": VIDEO_WITH_AUDIO_FILENAME
+                    "S3Bucket": testing_env_variables['BUCKET_NAME'],
+                    "S3Key": testing_env_variables['SAMPLE_VIDEO']
                     }
                 }
             }
         }
     
-    get_workflow_configuration_response = api.get_workflow_configuration_request(body, stack_resources)
+    get_workflow_configuration_response = api.get_workflow_configuration_request(body)
 
     configuration = get_workflow_configuration_response.json()
     print("Default Workflow Configuration = {}".format(configuration))
 
-    headers = {"Content-Type": "application/json", "Authorization": token}
+    headers = {"Content-Type": "application/json", "Authorization": testing_env_variables['token']}
     start_request = requests.post(stack_resources["WorkflowApiEndpoint"]+'/workflow/execution', headers=headers, json=body, verify=False)
     assert start_request.status_code == 200
     assert start_request.json()['Status'] == 'Queued'
     
     workflow_execution = start_request.json()
     
-    workflow_execution = api.wait_for_workflow_execution(workflow_execution, stack_resources, 240)
+    workflow_execution = api.wait_for_workflow_execution(workflow_execution, 600)
     assert workflow_execution["Status"] == "Complete"
 
     asset_id = workflow_execution["AssetId"]
@@ -100,8 +87,7 @@ def test_instant_translate(stack_resources, api_schema):
     for operator_name in ["Transcribe", "Translate"]:
         operator = {"OperatorName" : operator_name}
 
-        retrieve_single_metadata_response = api.get_single_metadata_field(stack_resources, asset_id,
-                                                                        operator)
+        retrieve_single_metadata_response = api.get_single_metadata_field(asset_id, operator)
         assert retrieve_single_metadata_response.status_code == 200
 
         retrieved_single_metadata = retrieve_single_metadata_response.json()
