--- tests/tests-parameterized-rekognition/requirements.txt
@@ -1,6 +1,0 @@
-boto3==1.9.139
-botocore==1.12.139
-pytest
-requests
-urllib3
-jsonschema
