--- tests/tests-workflowapi/requirements.txt
@@ -1,6 +1,0 @@
-boto3==1.9.139
-botocore==1.12.139
-pytest==4.5.0
-requests==2.22.0
-urllib3==1.24.2
-jsonschema==2.6.0
