+++ tests/xtest_workflow_execution_crud.py
@@ -29,21 +29,11 @@ import os
 from jsonschema import validate
 
 # local imports
-import api 
 import validation
 
-REGION = os.environ['REGION']
-BUCKET_NAME = os.environ['BUCKET_NAME']
-MIE_STACK_NAME = os.environ['MIE_STACK_NAME']
-VIDEO_FILENAME = os.environ['VIDEO_FILENAME']
-IMAGE_FILENAME = os.environ['IMAGE_FILENAME']
-AUDIO_FILENAME = os.environ['AUDIO_FILENAME']
-TEXT_FILENAME = os.environ['TEXT_FILENAME']
 
-
-
-def test_workflow_execution_api(operations, stages, workflow_configs, stack_resources, api_schema):
-
+def test_workflow_execution_api(api, api_schema, workflow_configs):
+    api = api()
     urllib3.disable_warnings(urllib3.exceptions.InsecureRequestWarning)
 
     print("Running /workflow/execution API tests")
@@ -53,26 +43,24 @@ def test_workflow_execution_api(operations, stages, workflow_configs, stack_reso
         print("\nTEST WORKFLOW EXECUTION CONFIGURATION: {}".format(config))
 
         # Create the workflow
-        create_workflow_response = api.create_workflow_request(config, stack_resources)
+        # TODO: Add a check here to see if the workflow is in a deleting state (from a previous test run) and wait 
+        create_workflow_response = api.create_workflow_request(config)
         workflow = create_workflow_response.json()
-        
-        assert create_workflow_response.status_code == 200
-        validation.schema(workflow, api_schema["create_workflow_response"])
 
-        get_workflow_response = api.get_workflow_request(workflow, stack_resources)
-        workflow = get_workflow_response.json()
-        assert get_workflow_response.status_code == 200
-        validation.schema(workflow, api_schema["create_workflow_response"])
+        assert create_workflow_response.status_code == 200
+        
+        #FIXME - validate create_workflow_response
+        #validation.schema(workflow, api_schema["create_workflow_response"])
 
         # Execute the workflow and wait for it to complete
-        create_workflow_execution_response = api.create_workflow_execution_request(workflow, config, stack_resources)
+        create_workflow_execution_response = api.create_workflow_execution_request(workflow, config)
         workflow_execution = create_workflow_execution_response.json()
         assert create_workflow_execution_response.status_code == 200
         assert workflow_execution['Status'] == 'Queued'
         
         #FIXME - validate create_workflow_response
         #validation.schema(workflow_execution, api_schema["create_workflow_execution_response"])
-        workflow_execution = api.wait_for_workflow_execution(workflow_execution, stack_resources, 120)
+        workflow_execution = api.wait_for_workflow_execution(workflow_execution, 120)
         if config["Status"] == "OK":
             assert workflow_execution["Status"] == "Complete"
             
@@ -87,11 +75,11 @@ def test_workflow_execution_api(operations, stages, workflow_configs, stack_reso
         # validation.stage_execution(workflow_execution, config, stack_resources, api_schema)
 
         # Delete the workflow
-        delete_workflow_response = api.delete_stage_workflow_request(workflow, stack_resources)
+        delete_workflow_response = api.delete_stage_workflow_request(workflow)
         assert delete_workflow_response.status_code == 200
         
         #Delete the workflow
-        delete_stage_response = api.delete_stage_request(workflow, stack_resources)
+        delete_stage_response = api.delete_stage_request(workflow)
         assert delete_stage_response.status_code == 200
